Add retry functionality
This commit adds a retry functionality to the Scrollbar-Fixer script. A counter variable and an extratries variable have been added to keep track of the number of times the fixScrollbars function has been called. If the scrollbars have not been fixed after the first timeout, the function will be called again timeout. The number of extra tries can be adjusted by changing the value of the extratries variable. This allows the script to handle websites where a longer timeout is needed to fix the scrollbars.