Add @exclude directive
This commit adds an @exclude directive to the script's metadata block to prevent it from running on specific websites. The @exclude directive specifies a URL or URL pattern that the script should not run on. In this case, the script has been updated to exclude the website https://github.com/.

Additionally, a comment has been added to the script to provide instructions on how to add more @exclude directives. Users can follow these instructions to prevent the script from running on additional websites as needed.

This update gives users more control over where the script runs and allows them to easily exclude specific websites if desired with clear detailed instructions.